% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_influence.R
\name{plot_multilevel_influence}
\alias{plot_multilevel_influence}
\title{Plot Multilevel Influence Diagnostics}
\usage{
plot_multilevel_influence(x, type = c("cook", "sigma"))
}
\arguments{
\item{x}{multilevel_influence object.}

\item{type}{"cook" or "sigma".}
}
\value{
ggplot object.
}
\description{
Plot Multilevel Influence Diagnostics
}
