% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_influence.R
\name{print.multilevel_influence}
\alias{print.multilevel_influence}
\title{Print Multilevel Influence Diagnostics}
\usage{
\method{print}{multilevel_influence}(x, ...)
}
\arguments{
\item{x}{multilevel_influence object.}

\item{...}{Additional arguments.}
}
\description{
Print Multilevel Influence Diagnostics
}
